chore: Remove old TomeType
